Separated another classes. Now the class decompozition is complete!
